+++ src/index.js
@@ -1,9 +1,19 @@
-import { entries, fromMap, get, keys, mergePaths, set, typeOf } from "./util";
+import {
+  entries,
+  fromMap,
+  get,
+  getPath,
+  keys,
+  mergePaths,
+  set,
+  typeOf,
+} from "./util";
 
 export const util = {
   entries,
   fromMap,
   get,
+  getPath,
   keys,
   mergePaths,
   set,
@@ -21,3 +31,4 @@ export { getPred } from "./pred";
 export { isOpt, opt, select } from "./selection";
 export { getSpread, spread } from "./spread";
 export { validate } from "./validate";
+export { validatePred } from "./validation";